refactor: extract vertical-layout core CSS to styles folder
