+++ packages/vertical-layout/src/vaadin-vertical-layout.js
@@ -9,7 +9,7 @@ import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { verticalLayoutStyles } from './vaadin-vertical-layout-styles.js';
+import { verticalLayoutStyles } from './styles/vaadin-vertical-layout-styles.js';
 
 /**
  * `<vaadin-vertical-layout>` provides a simple way to vertically align your HTML elements.
